html/semantics/forms/the-select-element/common-HTMLOptionsCollection.html

<!DOCTYPE html>
<html>
<head>
	<meta charset="utf-8">

	<title id='title'>HTMLOptionsCollection 2.8.3.2</title>
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>	

	<div id="log"></div>
    <select id="selly">
        <option>1</option>
        <option>2</option>
        <option>3</option>
        <option>4</option>
    </select>

    	<script>
    	    test(function () {
    	        /* Responding to notes from Ms2ger
                Mutation events removed, other items fixed. Would it be better to do a separate async test with mutation observers to cover requirement that mutation events fire?
                */
    	        var selly = document.getElementById('selly');

    	        test(function () {
    	            assert_equals(selly.length, 4);
    	        }, "On getting, the length attribute must return the number of nodes represented by the collection.");

    	        selly.length = 7;
    	        test(function () {
    	            assert_equals(selly.length, 7);
    	        }, "Changing the length adds new nodes; The number of new nodes = new length minus old length");

    	        var elarray = [];
    	        for (var i = 0; i < selly.length; i++) {
    	            elarray.push(selly[i].value);
    	        }

    	        test(function () {
    	            assert_array_equals(elarray, ["1", "2", "3", "4", "", "", ""]);
    	        }, "New nodes have no value");

    	        selly.length = 7;
    	        test(function () {
    	            assert_equals(selly.length, 7);
    	        }, "Setting a length equal to existing length changes nothing");


    	        selly.length = 4;
    	        test(function () {
    	            assert_equals(selly[6], undefined);
    	        }, "Setting a length lower than the old length trims nodes from the end (checking that previously set node is now undefined)");

    	        test(function () {
    	            assert_equals(selly.length, 4);
    	        }, "Number of nodes in collection is correctly changed");

    	    });
	</script>
